module4-task1 | move files js
